requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiosqlite
aiomysql
pydantic-settings
python-dotenv
google-genai
python-multipart
